refactor main.rb
