+++ Gemfile
@@ -1,3 +1,5 @@
 source 'https://rubygems.org'
 
+gem 'rubocop', require: false
 gem 'rubocop', '>= 1.0', '< 2.0'
+gem "neovim"
